Embedded images and download links in chat.
When there is a log of image or file, chat.js sends a post request to
tell Web Server to retrieve the file. Then, the images/files can be
accessed through a get request.

However, a web client may get the file before retrieval is done, so we
have to reload those not found images. This is yet to be implemented.